refactor(rust): make the auto-retry an implementation detail of repositories
