+++ implementations/rust/ockam/ockam_abac/src/policy/storage/resource_policy_repository.rs
@@ -3,6 +3,10 @@ use ockam_core::async_trait;
 use ockam_core::compat::boxed::Box;
 use ockam_core::compat::vec::Vec;
 use ockam_core::Result;
+#[cfg(feature = "std")]
+use ockam_node::database::AutoRetry;
+#[cfg(feature = "std")]
+use ockam_node::retry;
 
 /// This repository stores policies for resources.
 /// A policy is an expression which can be evaluated against an environment (a list of attribute
@@ -35,4 +39,40 @@ pub trait ResourcePoliciesRepository: Send + Sync + 'static {
 
     /// Delete the policy associated to a given resource name and action
     async fn delete_policy(&self, resource_name: &ResourceName, action: &Action) -> Result<()>;
+}
+
+#[cfg(feature = "std")]
+#[async_trait]
+impl<T: ResourcePoliciesRepository> ResourcePoliciesRepository for AutoRetry<T> {
+    async fn store_policy(
+        &self,
+        resource_name: &ResourceName,
+        action: &Action,
+        expression: &Expr,
+    ) -> Result<()> {
+        retry!(self.wrapped.store_policy(resource_name, action, expression))
+    }
+
+    async fn get_policy(
+        &self,
+        resource_name: &ResourceName,
+        action: &Action,
+    ) -> Result<Option<ResourcePolicy>> {
+        retry!(self.wrapped.get_policy(resource_name, action))
+    }
+
+    async fn get_policies(&self) -> Result<Vec<ResourcePolicy>> {
+        retry!(self.wrapped.get_policies())
+    }
+
+    async fn get_policies_by_resource_name(
+        &self,
+        resource_name: &ResourceName,
+    ) -> Result<Vec<ResourcePolicy>> {
+        retry!(self.wrapped.get_policies_by_resource_name(resource_name))
+    }
+
+    async fn delete_policy(&self, resource_name: &ResourceName, action: &Action) -> Result<()> {
+        retry!(self.wrapped.delete_policy(resource_name, action))
+    }
 }
