+++ implementations/rust/ockam/ockam_abac/src/policy/storage/resource_repository_sql.rs
@@ -3,13 +3,14 @@ use sqlx::encode::IsNull;
 use sqlx::error::BoxDynError;
 use sqlx::postgres::any::AnyArgumentBuffer;
 use sqlx::*;
+use std::sync::Arc;
 use tracing::debug;
 
+use crate::{Resource, ResourceName, ResourceType, ResourcesRepository};
 use ockam_core::async_trait;
 use ockam_core::Result;
+use ockam_node::database::AutoRetry;
 use ockam_node::database::{FromSqlxError, SqlxDatabase, ToVoid};
-
-use crate::{Resource, ResourceName, ResourceType, ResourcesRepository};
 
 #[derive(Clone)]
 pub struct ResourcesSqlxDatabase {
@@ -24,6 +25,18 @@ impl ResourcesSqlxDatabase {
         Self {
             database,
             node_name: node_name.to_string(),
+        }
+    }
+
+    /// Create a repository
+    pub fn make_repository(
+        database: SqlxDatabase,
+        node_name: &str,
+    ) -> Arc<dyn ResourcesRepository> {
+        if database.needs_retry() {
+            Arc::new(AutoRetry::new(Self::new(database, node_name)))
+        } else {
+            Arc::new(Self::new(database, node_name))
         }
     }
