+++ implementations/rust/ockam/ockam_api/src/authenticator/storage/authority_members_repository.rs
@@ -4,6 +4,8 @@ use ockam_core::async_trait;
 use ockam_core::compat::boxed::Box;
 use ockam_core::compat::vec::Vec;
 use ockam_core::Result;
+use ockam_node::database::AutoRetry;
+use ockam_node::retry;
 
 /// This repository stores project members on the Authority node
 #[async_trait]
@@ -25,4 +27,32 @@ pub trait AuthorityMembersRepository: Send + Sync + 'static {
         &self,
         pre_trusted_identities: &PreTrustedIdentities,
     ) -> Result<()>;
+}
+
+#[async_trait]
+impl<T: AuthorityMembersRepository> AuthorityMembersRepository for AutoRetry<T> {
+    async fn get_member(&self, identifier: &Identifier) -> Result<Option<AuthorityMember>> {
+        retry!(self.wrapped.get_member(identifier))
+    }
+
+    async fn get_members(&self) -> Result<Vec<AuthorityMember>> {
+        retry!(self.wrapped.get_members())
+    }
+
+    async fn delete_member(&self, identifier: &Identifier) -> Result<()> {
+        retry!(self.wrapped.delete_member(identifier))
+    }
+
+    async fn add_member(&self, member: AuthorityMember) -> Result<()> {
+        retry!(self.wrapped.add_member(member.clone()))
+    }
+
+    async fn bootstrap_pre_trusted_members(
+        &self,
+        pre_trusted_identities: &PreTrustedIdentities,
+    ) -> Result<()> {
+        retry!(self
+            .wrapped
+            .bootstrap_pre_trusted_members(pre_trusted_identities))
+    }
 }
