+++ implementations/rust/ockam/ockam_api/src/cli_state/identities_attributes.rs
@@ -1,4 +1,3 @@
-use crate::cli_state::AutoRetry;
 use crate::CliState;
 use ockam::identity::{
     IdentitiesAttributes, IdentityAttributesRepository, IdentityAttributesSqlxDatabase,
@@ -19,9 +18,6 @@ impl CliState {
         &self,
         node_name: &str,
     ) -> Arc<dyn IdentityAttributesRepository> {
-        Arc::new(AutoRetry::new(IdentityAttributesSqlxDatabase::new(
-            self.database(),
-            node_name,
-        )))
+        IdentityAttributesSqlxDatabase::make_repository(self.database(), node_name)
     }
 }
