+++ implementations/rust/ockam/ockam_api/src/cli_state/policies.rs
@@ -1,18 +1,11 @@
-use crate::cli_state::{AutoRetry, CliState};
+use crate::cli_state::CliState;
 use ockam_abac::{Policies, ResourcePolicySqlxDatabase, ResourceTypePolicySqlxDatabase};
-use std::sync::Arc;
 
 impl CliState {
     pub fn policies(&self, node_name: &str) -> Policies {
         Policies::new(
-            Arc::new(AutoRetry::new(ResourcePolicySqlxDatabase::new(
-                self.database(),
-                node_name,
-            ))),
-            Arc::new(AutoRetry::new(ResourceTypePolicySqlxDatabase::new(
-                self.database(),
-                node_name,
-            ))),
+            ResourcePolicySqlxDatabase::make_repository(self.database(), node_name),
+            ResourceTypePolicySqlxDatabase::make_repository(self.database(), node_name),
         )
     }
 }
