+++ implementations/rust/ockam/ockam_api/src/cli_state/repositories.rs
@@ -4,7 +4,6 @@ use ockam::identity::{
     CredentialSqlxDatabase,
 };
 use ockam_core::compat::sync::Arc;
-use ockam_node::database::DatabaseConfiguration;
 use ockam_vault::storage::{SecretsRepository, SecretsSqlxDatabase};
 
 use crate::cli_state::storage::*;
@@ -18,171 +17,54 @@ use crate::cli_state::{UsersRepository, UsersSqlxDatabase};
 /// stored in the database
 impl CliState {
     pub fn change_history_repository(&self) -> Arc<dyn ChangeHistoryRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => {
-                Arc::new(AutoRetry::new(ChangeHistorySqlxDatabase::new(database)))
-            }
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(ChangeHistorySqlxDatabase::new(database))
-            }
-        }
+        ChangeHistorySqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn identities_repository(&self) -> Arc<dyn IdentitiesRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => {
-                Arc::new(AutoRetry::new(IdentitiesSqlxDatabase::new(self.database())))
-            }
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(IdentitiesSqlxDatabase::new(self.database()))
-            }
-        }
+        IdentitiesSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn purpose_keys_repository(&self) -> Arc<dyn PurposeKeysRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => Arc::new(AutoRetry::new(
-                PurposeKeysSqlxDatabase::new(self.database()),
-            )),
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(PurposeKeysSqlxDatabase::new(self.database()))
-            }
-        }
+        PurposeKeysSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn secrets_repository(&self) -> Arc<dyn SecretsRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => {
-                Arc::new(AutoRetry::new(SecretsSqlxDatabase::new(self.database())))
-            }
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(SecretsSqlxDatabase::new(self.database()))
-            }
-        }
+        SecretsSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn vaults_repository(&self) -> Arc<dyn VaultsRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => {
-                Arc::new(AutoRetry::new(VaultsSqlxDatabase::new(self.database())))
-            }
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(VaultsSqlxDatabase::new(self.database()))
-            }
-        }
+        VaultsSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn enrollment_repository(&self) -> Arc<dyn EnrollmentsRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => Arc::new(AutoRetry::new(
-                EnrollmentsSqlxDatabase::new(self.database()),
-            )),
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(EnrollmentsSqlxDatabase::new(self.database()))
-            }
-        }
+        EnrollmentsSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn nodes_repository(&self) -> Arc<dyn NodesRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => {
-                Arc::new(AutoRetry::new(NodesSqlxDatabase::new(self.database())))
-            }
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(NodesSqlxDatabase::new(self.database()))
-            }
-        }
+        NodesSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn tcp_portals_repository(&self) -> Arc<dyn TcpPortalsRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => {
-                Arc::new(AutoRetry::new(TcpPortalsSqlxDatabase::new(self.database())))
-            }
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(TcpPortalsSqlxDatabase::new(self.database()))
-            }
-        }
+        TcpPortalsSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn projects_repository(&self) -> Arc<dyn ProjectsRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => {
-                Arc::new(AutoRetry::new(ProjectsSqlxDatabase::new(self.database())))
-            }
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(ProjectsSqlxDatabase::new(self.database()))
-            }
-        }
+        ProjectsSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn spaces_repository(&self) -> Arc<dyn SpacesRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => {
-                Arc::new(AutoRetry::new(SpacesSqlxDatabase::new(self.database())))
-            }
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(SpacesSqlxDatabase::new(self.database()))
-            }
-        }
+        SpacesSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn users_repository(&self) -> Arc<dyn UsersRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => {
-                Arc::new(AutoRetry::new(UsersSqlxDatabase::new(self.database())))
-            }
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(UsersSqlxDatabase::new(self.database()))
-            }
-        }
+        UsersSqlxDatabase::make_repository(self.database())
     }
 
     pub(super) fn user_journey_repository(&self) -> Arc<dyn JourneysRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => Arc::new(AutoRetry::new(
-                JourneysSqlxDatabase::new(self.application_database()),
-            )),
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(JourneysSqlxDatabase::new(self.application_database()))
-            }
-        }
+        JourneysSqlxDatabase::make_repository(self.application_database())
     }
 
     pub fn cached_credentials_repository(&self, node_name: &str) -> Arc<dyn CredentialRepository> {
-        let database = self.database();
-        match database.configuration {
-            DatabaseConfiguration::SqlitePersistent { .. }
-            | DatabaseConfiguration::SqliteInMemory { .. } => Arc::new(AutoRetry::new(
-                CredentialSqlxDatabase::new(self.database(), node_name),
-            )),
-            DatabaseConfiguration::Postgres { .. } => {
-                Arc::new(CredentialSqlxDatabase::new(self.database(), node_name))
-            }
-        }
+        CredentialSqlxDatabase::make_repository(self.database(), node_name)
     }
 }
