+++ implementations/rust/ockam/ockam_api/src/cli_state/resources.rs
@@ -1,13 +1,11 @@
-use crate::cli_state::AutoRetry;
 use crate::CliState;
 use ockam_abac::{Resources, ResourcesSqlxDatabase};
-use std::sync::Arc;
 
 impl CliState {
     pub fn resources(&self, node_name: &str) -> Resources {
-        Resources::new(Arc::new(AutoRetry::new(ResourcesSqlxDatabase::new(
+        Resources::new(ResourcesSqlxDatabase::make_repository(
             self.database(),
             node_name,
-        ))))
+        ))
     }
 }
