+++ implementations/rust/ockam/ockam_api/src/cli_state/storage/enrollments_repository.rs
@@ -1,9 +1,10 @@
+use crate::cli_state::enrollments::IdentityEnrollment;
+use crate::cloud::email_address::EmailAddress;
 use ockam::identity::Identifier;
 use ockam_core::async_trait;
 use ockam_core::Result;
-
-use crate::cli_state::enrollments::IdentityEnrollment;
-use crate::cloud::email_address::EmailAddress;
+use ockam_node::database::AutoRetry;
+use ockam_node::retry;
 
 /// This trait stores the enrollment status for local identities
 /// If an identity has been enrolled it is possible to retrieve:
@@ -29,4 +30,27 @@ pub trait EnrollmentsRepository: Send + Sync + 'static {
 
     /// Return true if the identity with the given name is enrolled
     async fn is_identity_enrolled(&self, name: &str) -> Result<bool>;
+}
+
+#[async_trait]
+impl<T: EnrollmentsRepository> EnrollmentsRepository for AutoRetry<T> {
+    async fn set_as_enrolled(&self, identifier: &Identifier, email: &EmailAddress) -> Result<()> {
+        retry!(self.wrapped.set_as_enrolled(identifier, email))
+    }
+
+    async fn get_enrolled_identities(&self) -> Result<Vec<IdentityEnrollment>> {
+        retry!(self.wrapped.get_enrolled_identities())
+    }
+
+    async fn get_all_identities_enrollments(&self) -> Result<Vec<IdentityEnrollment>> {
+        retry!(self.wrapped.get_all_identities_enrollments())
+    }
+
+    async fn is_default_identity_enrolled(&self) -> Result<bool> {
+        retry!(self.wrapped.is_default_identity_enrolled())
+    }
+
+    async fn is_identity_enrolled(&self, name: &str) -> Result<bool> {
+        retry!(self.wrapped.is_identity_enrolled(name))
+    }
 }
