+++ implementations/rust/ockam/ockam_api/src/cli_state/storage/spaces_repository.rs
@@ -1,6 +1,8 @@
 use crate::cloud::space::Space;
 use ockam_core::async_trait;
 use ockam_core::Result;
+use ockam_node::database::AutoRetry;
+use ockam_node::retry;
 
 /// This trait supports the storage of spaces as retrieved from the Controller
 ///
@@ -29,4 +31,35 @@ pub trait SpacesRepository: Send + Sync + 'static {
 
     /// Delete a space
     async fn delete_space(&self, space_id: &str) -> Result<()>;
+}
+
+#[async_trait]
+impl<T: SpacesRepository> SpacesRepository for AutoRetry<T> {
+    async fn store_space(&self, space: &Space) -> Result<()> {
+        retry!(self.wrapped.store_space(space))
+    }
+
+    async fn get_space(&self, space_id: &str) -> Result<Option<Space>> {
+        retry!(self.wrapped.get_space(space_id))
+    }
+
+    async fn get_space_by_name(&self, name: &str) -> Result<Option<Space>> {
+        retry!(self.wrapped.get_space_by_name(name))
+    }
+
+    async fn get_spaces(&self) -> Result<Vec<Space>> {
+        retry!(self.wrapped.get_spaces())
+    }
+
+    async fn get_default_space(&self) -> Result<Option<Space>> {
+        retry!(self.wrapped.get_default_space())
+    }
+
+    async fn set_default_space(&self, space_id: &str) -> Result<()> {
+        retry!(self.wrapped.set_default_space(space_id))
+    }
+
+    async fn delete_space(&self, space_id: &str) -> Result<()> {
+        retry!(self.wrapped.delete_space(space_id))
+    }
 }
