+++ implementations/rust/ockam/ockam_api/src/cli_state/storage/tcp_portals_repository.rs
@@ -2,6 +2,8 @@ use crate::nodes::models::portal::OutletStatus;
 use ockam_core::Result;
 use ockam_core::{async_trait, Address};
 use ockam_multiaddr::MultiAddr;
+use ockam_node::database::AutoRetry;
+use ockam_node::retry;
 use std::net::SocketAddr;
 
 /// The TcpPortalsRepository is responsible for accessing the configured tcp inlets and tcp outlets
@@ -30,6 +32,41 @@ pub trait TcpPortalsRepository: Send + Sync + 'static {
 
     /// Delete the configuration of a TcpOutlet for a given node name and worker address
     async fn delete_tcp_outlet(&self, node_name: &str, worker_addr: &Address) -> Result<()>;
+}
+
+#[async_trait]
+impl<T: TcpPortalsRepository> TcpPortalsRepository for AutoRetry<T> {
+    async fn store_tcp_inlet(&self, node_name: &str, tcp_inlet: &TcpInlet) -> Result<()> {
+        retry!(self.wrapped.store_tcp_inlet(node_name, tcp_inlet))
+    }
+
+    async fn get_tcp_inlet(&self, node_name: &str, alias: &str) -> Result<Option<TcpInlet>> {
+        retry!(self.wrapped.get_tcp_inlet(node_name, alias))
+    }
+
+    async fn delete_tcp_inlet(&self, node_name: &str, alias: &str) -> Result<()> {
+        retry!(self.wrapped.delete_tcp_inlet(node_name, alias))
+    }
+
+    async fn store_tcp_outlet(
+        &self,
+        node_name: &str,
+        tcp_outlet_status: &OutletStatus,
+    ) -> Result<()> {
+        retry!(self.wrapped.store_tcp_outlet(node_name, tcp_outlet_status))
+    }
+
+    async fn get_tcp_outlet(
+        &self,
+        node_name: &str,
+        worker_addr: &Address,
+    ) -> Result<Option<OutletStatus>> {
+        retry!(self.wrapped.get_tcp_outlet(node_name, worker_addr))
+    }
+
+    async fn delete_tcp_outlet(&self, node_name: &str, worker_addr: &Address) -> Result<()> {
+        retry!(self.wrapped.delete_tcp_outlet(node_name, worker_addr))
+    }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
