+++ implementations/rust/ockam/ockam_app_lib/Cargo.toml
@@ -41,6 +41,7 @@ ockam = { path = "../ockam", version = "^0.144.0", features = ["software_vault"]
 ockam_api = { path = "../ockam_api", version = "0.87.0", default-features = false, features = ["std"] }
 ockam_core = { path = "../ockam_core", version = "^0.122.0" }
 ockam_multiaddr = { path = "../ockam_multiaddr", version = "0.66.0", features = ["cbor", "serde"] }
+ockam_node = { path = "../ockam_node", version = "^0.134.0" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 sqlx = { version = "0.8.2", default-features = false }
