+++ implementations/rust/ockam/ockam_identity/src/purpose_keys/storage/purpose_keys_repository.rs
@@ -1,9 +1,12 @@
-use ockam_core::async_trait;
-use ockam_core::compat::boxed::Box;
-use ockam_core::Result;
-
 use crate::models::{Identifier, PurposeKeyAttestation};
 use crate::Purpose;
+use async_trait::async_trait;
+use ockam_core::compat::boxed::Box;
+use ockam_core::Result;
+#[cfg(feature = "std")]
+use ockam_node::database::AutoRetry;
+#[cfg(feature = "std")]
+use ockam_node::retry;
 
 // TODO: Only one PurposeKey per Purpose per Identity is supported for now
 
@@ -33,4 +36,35 @@ pub trait PurposeKeysRepository: Send + Sync + 'static {
 
     /// Delete all keys
     async fn delete_all(&self) -> Result<()>;
+}
+
+#[cfg(feature = "std")]
+#[async_trait]
+impl<T: PurposeKeysRepository> PurposeKeysRepository for AutoRetry<T> {
+    async fn set_purpose_key(
+        &self,
+        subject: &Identifier,
+        purpose: Purpose,
+        purpose_key_attestation: &PurposeKeyAttestation,
+    ) -> Result<()> {
+        retry!(self
+            .wrapped
+            .set_purpose_key(subject, purpose, purpose_key_attestation))
+    }
+
+    async fn delete_purpose_key(&self, subject: &Identifier, purpose: Purpose) -> Result<()> {
+        retry!(self.wrapped.delete_purpose_key(subject, purpose))
+    }
+
+    async fn get_purpose_key(
+        &self,
+        identifier: &Identifier,
+        purpose: Purpose,
+    ) -> Result<Option<PurposeKeyAttestation>> {
+        retry!(self.wrapped.get_purpose_key(identifier, purpose))
+    }
+
+    async fn delete_all(&self) -> Result<()> {
+        retry!(self.wrapped.delete_all())
+    }
 }
