+++ implementations/rust/ockam/ockam_node/src/storage/database/mod.rs
@@ -1,8 +1,10 @@
+mod auto_retry;
 mod database_configuration;
 mod migrations;
 mod sqlx_database;
 mod sqlx_from_row_types;
 
+pub use auto_retry::*;
 pub use database_configuration::*;
 pub use migrations::*;
 pub use sqlx_database::*;
